names improvements
