+++ src/main/kotlin/com/demo/order/domain/port/CreditCardDataAccessPort.kt
@@ -1,0 +1,8 @@
+package com.demo.order.domain.port
+
+import com.demo.order.domain.model.Card
+import com.demo.order.domain.model.Customer
+
+interface CreditCardDataAccessPort {
+    fun findCardByCustomer(customer: Customer): Card?
+}
