+++ src/main/kotlin/com/demo/order/domain/port/CustomerDataAccessPort.kt
@@ -1,0 +1,7 @@
+package com.demo.order.domain.port
+
+import com.demo.order.domain.model.Customer
+
+interface CustomerDataAccessPort {
+    fun findByCpf(cpf: String): Customer?
+}
