+++ src/main/kotlin/com/demo/order/domain/port/OrderDataAccessPort.kt
@@ -1,0 +1,7 @@
+package com.demo.order.domain.port
+
+import com.demo.order.domain.model.Order
+
+interface OrderDataAccessPort {
+    fun save(order: Order): Order
+}
