+++ src/main/kotlin/com/demo/order/domain/port/ProductDataAccessPort.kt
@@ -1,0 +1,8 @@
+package com.demo.order.domain.port
+
+import com.demo.order.domain.model.Product
+
+interface ProductDataAccessPort {
+    fun findByCodes(products: List<String>): List<Product>?
+}
+
