+++ src/main/kotlin/com/demo/order/infra/inmemory/repository/CustomerInMemoryRepository.kt
@@ -1,0 +1,14 @@
+package com.demo.order.infra.inmemory.repository
+
+import com.demo.order.domain.model.Customer
+
+const val CPF = "84917644020"
+
+val CUSTOMER = Customer(
+    name = MASTERCARD_NAME,
+    cpf = CPF,
+    cards = CARDS
+)
+
+val CUSTOMERS = listOf(CUSTOMER)
+
