+++ src/main/kotlin/com/demo/order/infra/inmemory/repository/OrderInMemoryRepository.kt
@@ -1,0 +1,5 @@
+package com.demo.order.infra.inmemory.repository
+
+import com.demo.order.domain.model.Order
+
+val ORDERS = mutableListOf<Order>()
