+++ src/main/kotlin/com/demo/order/infra/inmemory/repository/ProductInMemoryRepository.kt
@@ -1,8 +1,6 @@
-package com.demo.order.infra.repository.memory
+package com.demo.order.infra.inmemory.repository
 
 import com.demo.order.domain.model.Product
-import com.demo.order.domain.port.IProductRepository
-import org.springframework.stereotype.Repository
 
 const val PRODUCT_EAN_7895824271025 = "7895824271025"
 const val PRODUCT_NAME_MAC = "Macbook Pro M1"
@@ -26,8 +24,3 @@ val PRODUCTS = listOf(
     PRODUCT_7898242582360
 )
 
-@Repository
-class ProductMemoryRepository : IProductRepository {
-    override fun findByCodes(products: List<String>): List<Product>? =
-        PRODUCTS
-}
